separate some "core" components into a pyviewerlib/core directory.
